Atualiza script para gerar templates para depender da app django
